Switch to Qt6-only, fix Qt6 API compatibility (remove setCodec, fix qsizetype)
Agent-Logs-Url: https://github.com/Zer0C00I/MacroSIP/sessions/0a80e8b2-9afa-4abf-81e1-5fc8e7025a5a

+++ src/dialogs/ShortcutsDialog.cpp
@@ -18,7 +18,7 @@ ShortcutsDialog::ShortcutsDialog(QWidget *parent)
 void ShortcutsDialog::setShortcuts(const QList<Shortcut> &shortcuts)
 {
     for (int row = 0; row < RowCount; ++row) {
-        const bool hasData = row < shortcuts.size();
+        const bool hasData = row < static_cast<int>(shortcuts.size());
 
         auto *labelEdit = qobject_cast<QLineEdit *>(
             m_table->cellWidget(row, ColLabel));
